feat: move turn into agent call to action in input bar
